hw16-data-rest/src/main/java/ru/otus/spring/psannikov/data/rest/services/GenreServiceImpl.java
package ru.otus.spring.psannikov.data.rest.services;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import ru.otus.spring.psannikov.data.rest.models.Genre;
import ru.otus.spring.psannikov.data.rest.repositories.GenreRepository;

import java.util.List;
import java.util.Optional;

@RequiredArgsConstructor
@Service
public class GenreServiceImpl implements GenreService {
    private final GenreRepository genreRepository;

    @Override
    public Optional<Genre> findById(long id) {
        return genreRepository.findById(id);
    }

    @Override
    public List<Genre> findAll() {
        return genreRepository.findAll();
    }
}
